feat: multiverse background verses and info modal
