+++ packages/fe/assets/scss/variables.scss
@@ -27,7 +27,7 @@ $cove: #7794C5;
 $contessa: #D67272;
 $asparagus: #88995F;
 $effy: #F2A60D;
-$woodsmoke: #0B0C0E;
+$woodsmoke: #292B33;
 $athensGray: #EFF0F3;
 $gullGray: #949DB8;
 $stormGray: #6B7080;
