+++ packages/fe/components/icon/close.vue
@@ -1,0 +1,7 @@
+<template>
+  <svg width="18" height="19" viewBox="0 0 18 19" fill="none" xmlns="http://www.w3.org/2000/svg">
+<rect x="16.3636" y="0.5" width="2.31416" height="23.1416" rx="1.15708" transform="rotate(45 16.3636 0.5)" fill="#8393C0"/>
+<path fill-rule="evenodd" clip-rule="evenodd" d="M7.36348 7.86361C6.91162 8.31548 6.17899 8.31548 5.72713 7.86361L0.818052 2.95454C0.366184 2.50267 0.366183 1.77005 0.818051 1.31818C1.26992 0.866312 2.00254 0.866311 2.45441 1.31818L7.36348 6.22725C7.81535 6.67912 7.81535 7.41174 7.36348 7.86361ZM10.6362 12.7727C10.1843 12.3208 10.1843 11.5882 10.6362 11.1363C11.0881 10.6845 11.8207 10.6845 12.2726 11.1363L17.1816 16.0454C17.6335 16.4973 17.6335 17.2299 17.1816 17.6818C16.7298 18.1336 15.9971 18.1336 15.5453 17.6818L10.6362 12.7727Z" fill="#8393C0"/>
+</svg>
+
+</template>
