+++ packages/fe/data/info.md
@@ -1,8 +1,8 @@
-## about fuit.es
+*fuit.es* aims to provide a creative online space, a platform for collective processes designed with the ‘Anarchive’ as its operational concept; an archive which promotes the continual evolution and variation of its documents rather than their mere preservation.
 
-Fuit.es aims to provide a creative online space, a platform for collective processes designed with the ‘Anarchive’ as its operational concept; an archive which promotes the continual evolution and variation of its documents rather than their mere preservation.
+Each individual page offers a space for collaboration, where digital media in the form of images, sounds or text can be uploaded and composed with. These individual media objects are referred to as “thingies”. Anyone with a token can create a verse, a public network of pages for anarchiving thingies. If you are curious about accessing the platform, get in touch with us by email at [contact@fuit.es](mailto:contact@fuit.es), we would be happy to make you a token.
 
-Each individual page offers a space for collaboration, where digital media in the form of images, sounds or text can be uploaded and composed with. These individual media objects are referred to as “thingies”. Anyone with a token can create a verse, a public network of pages for anarchiving thingies. If you are curious about accessing the platform, get in touch with us by email at contact@fuit.es, we would be happy to make you a token.
+---
 
-###### *We acknowledge the support of the Canada Council for the Arts.*
-###### *Nous remercions le Conseil des arts du Canada de son soutien.*
+###### *We acknowledge the support of the [Canada Council for the Arts](https://canadacouncil.ca/).*
+###### *Nous remercions le [Conseil des arts du Canada](https://canadacouncil.ca/) de son soutien.*
