+++ packages/fe/data/settings.json
@@ -210,11 +210,6 @@
   },
   "landingSites": [
     {
-      "label": "explore mode",
-      "slug": "explore",
-      "tooltip": "explore-mode-toggle"
-    },
-    {
       "label": "tooltips",
       "slug": "tooltips",
       "tooltip": "tooltip-mode-toggle"
@@ -244,5 +239,6 @@
       "to": "mailto:contact@fuit.es",
       "theme": "clear"
     }
-  ]
+  ],
+  "infoModalMarkdown": "## about fuit.es\n\nFuit.es aims to provide a creative online space, a platform for collective processes designed with the 'Anarchive' as its operational concept; an archive which promotes the continual evolution and variation of its documents rather than their mere preservation.\n\nEach individual page offers a space for collaboration, where digital media in the form of images, sounds or text can be uploaded and composed with. These individual media objects are referred to as “thingies”. Anyone with a token can create a verse, a public network of pages for anarchiving thingies. If you are curious about accessing the platform, get in touch with us by email at contact@fuit.es, we would be happy to make you a token.\n\n###### *We acknowledge the support of the Canada Council for the Arts.*\n###### *Nous remercions le Conseil des arts du Canada de son soutien.*"
 }
